--- barcodeokapi/lib/ui/count/home.dart
@@ -1,41 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-
-/// Annotating a class by `@riverpod` defines a new shared state for your application,
-/// accessible using the generated [counterProvider].
-/// This class is both responsible for initializing the state (through the [build] method)
-/// and exposing ways to modify it (cf [increment]).
-@riverpod
-class Counter extends _$Counter {
-  /// Classes annotated by `@riverpod` **must** define a [build] function.
-  /// This function is expected to return the initial state of your shared state.
-  /// It is totally acceptable for this function to return a [Future] or [Stream] if you need to.
-  /// You can also freely define parameters on this method.
-  @override
-  int build() => 0;
-
-  void increment() => state++;
-}
-
-class Home extends ConsumerWidget {
-  @override
-  Widget build(context, WidgetRef ref) {
-    return Scaffold(
-      appBar: AppBar(title: const Text('Counter example')),
-      body: Center(
-        // Consumer is a widget that allows you reading providers.
-        child: Consumer(
-          builder: (context, ref, _) {
-            final count = ref.watch(counterProvider);
-            return Text('$count');
-          },
-        ),
-      ),
-      floatingActionButton: FloatingActionButton(
-        // The read method is a utility to read a provider without listening to it
-        onPressed: () => ref.read(counterProvider.notifier).increment(),
-        child: const Icon(Icons.add),
-      ),
-    );
-  }
-}
