+++ barcodeokapi/pubspec.yaml
@@ -9,34 +9,28 @@ environment:
   sdk: '>=2.18.6 <3.0.0'
   flutter: ">=2.0.0"
 
+
 dependencies:
+  cupertino_icons: ^1.0.2
   flutter:
     sdk: flutter
-
-  # iOS デザイン
-  cupertino_icons:
-  # immutable な model作成用
-  freezed_annotation:
-  # 状態管理パッケージ
   flutter_riverpod:
-  json_annotation:
-
+  freezed_annotation: ^2.2.0
+  hooks_riverpod:
+  json_annotation: ^4.8.0
   riverpod:
   riverpod_annotation:
 
 dev_dependencies:
+  build_runner: ^2.3.3
+  # build_verify: ^3.0.0
+  # custom_lint: ^0.3.2
   flutter_test:
     sdk: flutter
-  flutter_lints:
-  # ビルドランナー, freezed 等の監視に使用
-  build_runner:
-  build_verify:
-  # immutable な model 作成用
-  freezed:
-  json_serializable:
+  freezed: ^2.3.2
+  json_serializable: ^6.6.1
   riverpod_generator:
-
-  # custom_lint: ^0.3.2
+  # riverpod_lint:
 
 flutter:
   uses-material-design: true
